Fix duplicate FormField
