+++ docs/src/chat/components/ChatInput.tsx
@@ -1,6 +1,6 @@
 import { Input } from "@base-ui/react";
 import { forwardRef } from "react";
-import { useField } from "../form/FormContext";
+import { useField } from "../form/FieldContext";
 
 interface ChatInputProps {
   placeholder?: string;
