+++ docs/src/lib/site-meta.ts
@@ -3,11 +3,13 @@ import type { Metadata } from "next";
 
 export const cliVersion = "1.2.0";
 export const projectName = "Worktree";
-export const projectDescription = "A CLI tool for managing git worktrees with enhanced workflow features";
+export const projectDescription =
+  "A CLI tool for managing git worktrees with enhanced workflow features";
 export const projectLink = "https://github.com/burglekitt/worktree";
 export const projectOwnerName = "burglekitt";
 export const projectOwnerProfileUrl = "https://github.com/burglekitt";
-export const projectOwnerAvatarUrl = "https://github.com/burglekitt.png?size=64";
+export const projectOwnerAvatarUrl =
+  "https://github.com/burglekitt.png?size=64";
 
 export const metadata: Metadata = {
   title: {
